fix(editor): Move project components
