+++ packages/editor-ui/src/components/Projects/ProjectCardBadge.vue
