+++ packages/editor-ui/src/components/Projects/ProjectNavigation.vue
