+++ packages/editor-ui/src/components/Projects/ProjectSharingInfo.vue
