+++ packages/editor-ui/src/components/Projects/ProjectTabs.vue
