+++ packages/editor-ui/src/components/forms/ResourceFiltersDropdown.vue
@@ -52,7 +52,7 @@ import { mapStores } from 'pinia';
 import { useProjectsStore } from '@/stores/projects.store';
 import type { PropType } from 'vue';
 import type { ProjectSharingData } from '@/types/projects.types';
-import ProjectSharing from '@/features/projects/components/ProjectSharing.vue';
+import ProjectSharing from '@/components/Projects/ProjectSharing.vue';
 
 export type IResourceFiltersType = Record<string, boolean | string | string[]>;
